+++ src/FullHealth/Configuration/WorkMode.cs
@@ -7,6 +7,7 @@ public enum WorkMode : byte
     Host = 1 << 0,
     Client = 1 << 1,
     All = Host | Client,
+
     [Obsolete]
     HostAndClient = All
 }