+++ src/FullHealth/Helpers/HealPlayerHelper.cs
@@ -6,7 +6,7 @@ public static class HealPlayerHelper
 {
     private static readonly Random Random = new();
 
-    public static void Heal(PlayerAvatar player)
+    public static void Heal(PlayerAvatar player, int playerCount)
     {
         if (player?.playerHealth == null)
         {
@@ -19,12 +19,24 @@ public static class HealPlayerHelper
             Plugin.Logger.LogDebug($"The player '{player.playerName}' did not survive and will not be healed");
             return;
         }
-        
+
         var playerHealthValue = player.playerHealth.health;
 
         int healValue;
 
-        if (Configuration.HealthPackModeValues is { Length: > 0 })
+        if (!string.IsNullOrWhiteSpace(Configuration.Expression.Value))
+        {
+            var result = ExpressionHelper.Calculate(player, playerCount);
+
+            if (double.IsNaN(result))
+            {
+                Plugin.Logger.LogError($"Error in calculating the heal value for player '{player.playerName}' using the expression. Check the expression.");
+                return;
+            }
+
+            healValue = Convert.ToInt32(result);
+        }
+        else if (Configuration.HealthPackModeValues is { Length: > 0 })
         {
             var index = Configuration.HealthPackModeValues.Length == 1 ? 0 : Random.Next(0, Configuration.HealthPackModeValues.Length);
             healValue = Configuration.HealthPackModeValues[index];
@@ -33,9 +45,13 @@ public static class HealPlayerHelper
         {
             healValue = Configuration.ExactValue.Value;
         }
+        else if (Configuration.ByMaxHealthPercentage.Value > 0)
+        {
+            healValue = decimal.ToInt32(player.playerHealth.maxHealth * (Configuration.ByMaxHealthPercentage.Value / 100));
+        }
         else
         {
-            var expectedHealth = decimal.ToInt32(player.playerHealth.maxHealth * (Configuration.Percent.Value / 100));
+            var expectedHealth = decimal.ToInt32(player.playerHealth.maxHealth * (Configuration.ToMaxHealthPercentage.Value / 100));
 
             if (playerHealthValue >= expectedHealth)
             {
@@ -50,7 +66,7 @@ public static class HealPlayerHelper
 
         player.playerHealth.HealOther(healValue, true);
 
-        Plugin.Logger.LogDebug($"Player '{player.playerName}' with '{playerHealthValue}' health received '{healValue}' HP. " +
+        Plugin.Logger.LogDebug($"Player '{player.playerName}' with '{playerHealthValue}' HP received '{healValue}' HP. " +
                                $"So player should have '{Math.Min(player.playerHealth.maxHealth, playerHealthValue + healValue)}' HP");
     }
 }