+++ src/FullHealth/Plugin.cs
@@ -1,6 +1,7 @@
 ﻿using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
+using org.mariuszgromada.math.mxparser;
 
 namespace FullHealth;
 
@@ -12,6 +13,7 @@ public class Plugin : BaseUnityPlugin
 
     private void Awake()
     {
+        License.iConfirmNonCommercialUse("yvsimakov");
         Logger = base.Logger;
         ConfigurationHelper.Bind(Config);
         var harmony = new Harmony(MyPluginInfo.PLUGIN_GUID);